+++ screens/lesson_screen.py
@@ -1,0 +1,7 @@
+from kivy.uix.screenmanager import Screen
+from kivy.uix.floatlayout import FloatLayout
+
+class LessonScreen(Screen):
+
+    class LessonScreenCanvas(FloatLayout):
+        pass
