+++ screens/lesson_screen_overview.py
@@ -1,0 +1,7 @@
+from kivy.uix.screenmanager import Screen
+from kivy.uix.floatlayout import FloatLayout
+
+class Lessonscreenoverview(Screen):
+
+    class LessonscreenoverviewCanvas(FloatLayout):
+        pass
